demandas-nuvem/app/backend/src/requirements.txt
PyJWT==2.3.0
PyJWT==2.10.1
pymongo==4.11.2
python_bcrypt==0.3.2
